src/job_image_infer/requirements.txt
torch==2.1.0
torchvision==0.16.0
pillow>=10.0.0
